countingCards_switchStatement
This was my first experiment with using a switch statement in JavaScript. This program helps the user determine how he should play a hand of blackjack using the card counting method. Disclaimer: counting cards outside of an academic setting is not recommended and could be hazardous to your well-being (but it can be fun!).