Palindrome Test
This algorithm takes a string and tests it to see if it is a palindrome.

What I learned:
How to combine multiple algorithms to create a larger function: I incorporated the reverse string algorithm I created earlier to reverse the given string value and stored it in a new variable. This allowed me to compare to original string to the reversed string. 

string.toLowerCase(): In JavaScript, "a" and "A" are two different characters. To ensure all palindromes are identified regardless of capitalization, I converted all characters to lowercase.  

RegExp Metacharacters: I applied the \s and \w RegEx metacharacters to remove all non-alphanumeric characters from the test cases. Doing so allowed the program to compare strings based only on its alphanumeric attributes. Doing so ensures that palindromes are identified regardless of how non-alphanumeric characters are placed in the string.