Find the Length of the Longest Word in a String
This program determines the length of the longest word in a given string.

What I Learned:
Converting a String to an Array: I used the string.split() property to convert the given string to an array. I set the property parameters to split the string at each space, ensuring each word becomes a separate array element.
Using a For Loop to Compare Array Elements: Using this loop ensured each array element was examined and compared to the longest array element while automatically terminating at the end of the string.