Capitalize the 1st Letter of Each Word in String
This algorithm takes the first letter of each word in a string and capitalize the first letter.

What I Learned:
string.charAt() method: This method selects a character at the designated number in an element. In this case, I used the method to select the first character in a string and modified it to an uppercase character.

string.substring() method: This method creates a substring from a string beginning with the character at the designated number in an element. In this case, I used this method to concatenate the uppercase letter with a substring beginning with the first lowercase letter of the array element. In other applications, this method could be used to create a substring by indicating a start and endpoint within a string element (ie - string.substring(x, y)).